pyproject.toml
[tool.poetry]
name = "shipstation-client"
version = "0.1.0"
description = 'Python bindings for the ShipStation API'
readme = "README.md"
authors = ["AgriTheory <hello@agritheory.com>"]
license = "MIT"
repository = "https://github.com/agritheory/shipstation-client"
homepage = "https://github.com/agritheory/shipstation-client"
keywords = ['shipstation']
packages = [{ include = "shipstation" }]
classifiers = [
  "Environment :: Web Environment",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python",
  "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
  "Topic :: Internet :: WWW/HTTP",
]

[build-system]
requires = ["poetry>=1.0"]
build-backend = "poetry.masonry.api"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
attrs = "^23.2.0"
cattrs = "^23.2.3"
httpx = "^0.27.0"
importlib_metadata = {version = "^4.6.0", python = "<3.10"}
python-dateutil = "^2.8.1"

[tool.poetry.group.dev.dependencies]
bandit = "^1.7.8"
black = "^24.4.2"
darglint = "^1.8.1"
isort = "^5.13.2"
mypy = "^1.10.0"
pre-commit = "^3.7.1"
pydocstyle = "^5.0.2"
pytest = "^8.2.1"
pytest-asyncio = "^0.23.7"
pytest-cov = "^5.0.0"
pytest-httpx = "^0.30.0"
pytest-mypy = "^0.10.3"
python-dotenv = "^0.13.0"
pyupgrade = "^3.15.2"
respx = "^0.21.1"
safety = "^3.2.0"

[tool.pytest.ini_options]
addopts = "--cov=shipstation --cov-report term"

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.isort]
default_section = "FIRSTPARTY"
force_grid_wrap = 0
include_trailing_comma = true
known_typing = "typing,types,typing_extensions,mypy,mypy_extensions"
line_length = 80
multi_line_output = 3
sections = "TYPING,STDLIB,THIRDPARTY,FIRSTPARTY,LOCALFOLDER"
use_parentheses = true
